requirements.txt
asgiref==3.5.0
Django==4.0.4
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
dnspython==2.2.1
mongoengine==0.24.1
pymongo==4.1.1
pytz==2022.1
sqlparse==0.4.2
